what

+++ ring.h
@@ -3,7 +3,9 @@
 
 #define our_k 5
 
-typedef __int128_t scal;
+#include <stdint.h>
+
+typedef int64_t scal;
 
 typedef scal ring[2];
 
@@ -13,7 +15,7 @@ void ring_sub(ring z, const ring x, const ring y);
 
 void ring_mul(ring z, const ring x, const ring y);
 
-void ring_scale(ring z, const ring x, int scale_1, int scale_2);
+void ring_scale(ring z, const ring x, int64_t scale_1, int64_t scale_2);
 
 int ring_sign(const scal x);
